+++ Magna/src/main/java/de/morigm/magna/api/gui/GuiButton.java
@@ -1,6 +1,26 @@
 package de.morigm.magna.api.gui;
 
+import org.bukkit.inventory.ItemStack;
+
+import lombok.Getter;
+
 public class GuiButton 
 {
+	
+	@Getter private final ItemStack item;
+	@Getter private final int id;
+	@Getter private final int slot;
+
+	public GuiButton(ItemStack item,int id)
+	{
+		this(item,id,-1);
+	}
+
+	public GuiButton(ItemStack item, int id, int slot) 
+	{
+		this.item = item;
+		this.id = id;
+		this.slot = slot;
+	}
 
 }