+++ Magna/src/main/java/de/morigm/magna/stuff/MagnaStuff.java
@@ -7,6 +7,7 @@ import java.util.Map;
 
 import org.bukkit.entity.Player;
 
+import de.morigm.magna.api.gui.Gui;
 import lombok.Getter;
 
 public class MagnaStuff 
@@ -20,4 +21,5 @@ public class MagnaStuff
 	
 	@Getter private static Map<String,String> msg;
 	@Getter private static List<Player> onlybreaks;
+	@Getter private static Map<Player,Gui> guis = new HashMap<>();
 }